+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
 build/
-src/video/
+src/video/*.mp4
