+++ package.json
@@ -5,7 +5,9 @@
   "main": "index.ts",
   "scripts": {
     "dev": "nodemon './server.ts'",
-    "build": "tsc -p './tsconfig.json'",
+    "build": "npm-run-all build:ts scripts:build",
+    "build:ts": "tsc -p './tsconfig.json'",
+    "scripts:build": "ts-node ./scripts/build.ts",
     "lint": "eslint .",
     "lint:fix": "eslint --fix ."
   },
@@ -27,6 +29,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.0.0",
     "nodemon": "^2.0.20",
+    "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
